refactor 'service user banner' into its own presenter class
